frugen: json: Fix loading 'null' areas
Although `frugen` would not crash after the previous commit
if a key for an area in the input JSON file was present, but
the data was specified as `null`, it still would fail with an
error because json-c's API succeeds on such objects, but the
returned object is a null pointer.

Add a check for null object before enabling an area and
silently skip the area as version 2.x did.

Signed-off-by: Alexander Amelkin <alexander@amelkin.msk.ru>