libfru: Fix a potential out-of-bounds acccess
When setting an internal area value from a hex string using
fru_set_internal_hexstring(), the memory allocated for the
new copy of the source string wasn't cleared and then the
copied string wasn't terminated.

If the allocated block would happen to contain non-zero bytes,
that would lead to fru_t.internal string to become unterminated,
which in turn would lead to an out-of-bounds access when processing
that string.

Found with valgrind and fixed.

Signed-off-by: Alexander Amelkin <alexander@amelkin.msk.ru>