frugen: Fix a bug in product.ver processing
Due to a bug, when exporting to JSON, product.ver field
was written as `version`. When importing from JSON,
the expected name was though `ver`, as specified in `example.json`.

That was due to differences in the import and export code.

Refactor the export code to reuse the same array of field names.

Signed-off-by: Alexander Amelkin <alexander@amelkin.msk.ru>